Introduce SchedulingContext
